requirements.txt
ctypes
PyYAML
smtplib
simple-chalk
concurrent.futures
